webauthn/createcredential-getpublickey.https.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>WebAuthn getPublicKey</title>
<meta name="timeout" content="long">
<link rel="help" href="https://w3c.github.io/webauthn/#sctn-public-key-easy">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="helpers.js"></script>
<script>
const t = async_test("getPublicKey");

standardSetup(function() {
    createCredential().then((cred) => {
        const response = cred.response;

        t.step(function() {
            assert_equals(response.getPublicKeyAlgorithm(),
                          cose_alg_ECDSA_w_SHA256);
            const authData = response.getAuthenticatorData();
            assert_greater_than(authData.byteLength, 10);
            const spki = new Uint8Array(response.getPublicKey());
            assert_greater_than(spki.byteLength, 10);
            // DER SEQUENCEs start with an 0x30 byte.
            assert_equals(spki[0], 0x30);
            t.done();
        });
    }, (e) => {
        t.step(function() {
            assert_unreached("creation failed");
            t.done();
        });
    });
});

/* JSHINT */
/* globals standardSetup, createCredential */
</script>
</head>
<body></body>
</html>
